+++ clang-tools-extra/test/clang-tidy/Inputs/absl/external-file.h
@@ -1,1 +1,6 @@
-namespace absl {}
+namespace absl {
+namespace base_internal {
+void InternalFunction() {}
+} // namespace base_internal 
+} //namespace absl
+void DirectAccess2() { absl::base_internal::InternalFunction(); }